MNT: Add support for osx_arm64 architecture in conda-forge.yml and update build script

+++ recipe/build.sh
@@ -6,7 +6,7 @@ mkdir -p $PREFIX/share
 cp -R code-server $PREFIX/share/
 
 # Remove faulty files
-if [[ "${target_platform}" != "linux-aarch64" ]]; then
+if [[ "${target_platform}" != "linux-aarch64" && "${target_platform}" != "osx-arm64" ]]; then
   find $PREFIX/share -name node_gyp_bins | xargs rm -r
 fi
